moved sharedpreferences to settingsClass object first to avoid repeat sys calls
had created a onsharedprefs listener in the counterfragment but it can't communicate effectively with the prefs fragment so i'll have to move it to the prefs fragment and communicate through the activity or move the prefs fragment to it's own activity. i dunno. have to redesign the whole thing in photoshop :(

my baby, i hope u turn out alright in the end